backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
passlib[bcrypt]==1.7.4
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.9
qrcode[pil]==7.4.2
Pillow==10.1.0
aiofiles==23.2.1
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
email-validator
passlib[bcrypt]
python-jose[cryptography]==3.3.0
asyncpg>=0.29.0
jinja2==3.1.3
google-cloud-storage==2.14.0
ruff

